server/requirements.txt
paho-mqtt==1.3.0
protobuf==3.4.0
